Some updates in system

+++ erp_cultos-app/database/migrations/2023_03_22_093938_create_departments_table.php
@@ -18,7 +18,7 @@ class CreateDepartmentsTable extends Migration
             $table->string('Name');
             
             //Inicio da coluna contendo a chave estrangeiraa
-            $table->double('Id_user')->unsigned();
+            $table->integer('Id_user')->unsigned();
             $table->foreign('Id_user')->references('id')->on('users');
 
             $table->timestamps();
